small cleanup
