+++ test.sh
@@ -3,4 +3,4 @@ set -o errexit
 
 ls tests/*.sh | parallel --halt now,fail=1 bash
 PYTHONPATH=/ingest-validation-tools
-python -m unittest tests/test-dataset-examples.py
+python -m unittest tests/test_dataset_examples.py
