+++ tests_manual/README.md
