+++ tests_manual/test-plugins.sh
